Use project id for project field in index document.
The project field was previously automatically populated with the name
of the directory relative to the crawler's root path. This is useful
when indexing projects that are not signac projects. However, for signac
projects one would expect that the project field is identical to the
projects configured name.

The field will now be populated with a signac project's configured id.